+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ## Unreleased
 
+Configuration:
+* Add `-optF=main:_` option for configuring the name of the main module/function ([#64](https://github.com/brandonchinn178/skeletest/issues/64))
+
 API changes:
 * Added `shouldReturn`